+++ R/factors.R
@@ -44,17 +44,17 @@ factor_ <- function(x = integer(), levels = NULL, order = TRUE,
   } else {
     lvls <- levels
   }
-  if (na_exclude && anyNA(lvls)){
+  if (na_exclude && any_na(lvls)){
     if (order && is.null(levels)){
       lvls <- lvls[seq_len(length(lvls) - 1L)]
     } else {
-      lvls <- lvls[cpp_which_not_na(lvls)]
+      lvls <- lvls[which_not_na(lvls)]
     }
   }
   out <- collapse::fmatch(x, lvls, overid = 2L)
   fct_lvls <- as.character(lvls)
   if (inherits(x, "POSIXt") && collapse::any_duplicated(fct_lvls)){
-    fct_lvls <- paste(fct_lvls, as.POSIXlt(x)$zone)
+    fct_lvls <- paste(fct_lvls, as.POSIXlt(lvls)$zone)
   }
   attr(out, "levels") <- fct_lvls
   class(out) <- c(if (ordered) "ordered" else character(),