+++ src/cpp11.cpp
@@ -47,63 +47,98 @@ extern "C" SEXP _cheapr_cpp_lcm2_vectorised(SEXP x, SEXP y, SEXP tol, SEXP na_rm
     return cpp11::as_sexp(cpp_lcm2_vectorised(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(y), cpp11::as_cpp<cpp11::decay_t<double>>(tol), cpp11::as_cpp<cpp11::decay_t<bool>>(na_rm)));
   END_CPP11
 }
-// na_utils.cpp
-SEXP cpp_row_na_counts(SEXP x);
-extern "C" SEXP _cheapr_cpp_row_na_counts(SEXP x) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(cpp_row_na_counts(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
-  END_CPP11
-}
-// na_utils.cpp
-SEXP cpp_missing_row(SEXP x, double threshold, bool threshold_is_prop);
-extern "C" SEXP _cheapr_cpp_missing_row(SEXP x, SEXP threshold, SEXP threshold_is_prop) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(cpp_missing_row(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<double>>(threshold), cpp11::as_cpp<cpp11::decay_t<bool>>(threshold_is_prop)));
-  END_CPP11
-}
-// na_utils.cpp
+// nas.cpp
 SEXP cpp_num_na(SEXP x);
 extern "C" SEXP _cheapr_cpp_num_na(SEXP x) {
   BEGIN_CPP11
     return cpp11::as_sexp(cpp_num_na(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
   END_CPP11
 }
-// na_utils.cpp
+// nas.cpp
+bool cpp_any_na(SEXP x);
+extern "C" SEXP _cheapr_cpp_any_na(SEXP x) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_any_na(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+  END_CPP11
+}
+// nas.cpp
+bool cpp_all_na(SEXP x, bool return_true_on_empty);
+extern "C" SEXP _cheapr_cpp_all_na(SEXP x, SEXP return_true_on_empty) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_all_na(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<bool>>(return_true_on_empty)));
+  END_CPP11
+}
+// nas.cpp
 SEXP cpp_which_na(SEXP x);
 extern "C" SEXP _cheapr_cpp_which_na(SEXP x) {
   BEGIN_CPP11
     return cpp11::as_sexp(cpp_which_na(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
   END_CPP11
 }
-// na_utils.cpp
+// nas.cpp
 SEXP cpp_which_not_na(SEXP x);
 extern "C" SEXP _cheapr_cpp_which_not_na(SEXP x) {
   BEGIN_CPP11
     return cpp11::as_sexp(cpp_which_not_na(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
   END_CPP11
 }
-// na_utils.cpp
+// nas.cpp
+SEXP cpp_is_na(SEXP x);
+extern "C" SEXP _cheapr_cpp_is_na(SEXP x) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_is_na(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+  END_CPP11
+}
+// nas.cpp
+SEXP cpp_row_na_counts(SEXP x);
+extern "C" SEXP _cheapr_cpp_row_na_counts(SEXP x) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_row_na_counts(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+  END_CPP11
+}
+// nas.cpp
+SEXP cpp_col_na_counts(SEXP x);
+extern "C" SEXP _cheapr_cpp_col_na_counts(SEXP x) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_col_na_counts(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+  END_CPP11
+}
+// nas.cpp
+SEXP cpp_missing_row(SEXP x, double threshold, bool threshold_is_prop);
+extern "C" SEXP _cheapr_cpp_missing_row(SEXP x, SEXP threshold, SEXP threshold_is_prop) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_missing_row(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<double>>(threshold), cpp11::as_cpp<cpp11::decay_t<bool>>(threshold_is_prop)));
+  END_CPP11
+}
+// nas.cpp
+SEXP cpp_missing_col(SEXP x, double threshold, bool threshold_is_prop);
+extern "C" SEXP _cheapr_cpp_missing_col(SEXP x, SEXP threshold, SEXP threshold_is_prop) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_missing_col(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<double>>(threshold), cpp11::as_cpp<cpp11::decay_t<bool>>(threshold_is_prop)));
+  END_CPP11
+}
+// nas.cpp
 SEXP cpp_matrix_row_na_counts(SEXP x);
 extern "C" SEXP _cheapr_cpp_matrix_row_na_counts(SEXP x) {
   BEGIN_CPP11
     return cpp11::as_sexp(cpp_matrix_row_na_counts(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
   END_CPP11
 }
-// na_utils.cpp
+// nas.cpp
 SEXP cpp_matrix_col_na_counts(SEXP x);
 extern "C" SEXP _cheapr_cpp_matrix_col_na_counts(SEXP x) {
   BEGIN_CPP11
     return cpp11::as_sexp(cpp_matrix_col_na_counts(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
   END_CPP11
 }
-// na_utils.cpp
+// nas.cpp
 SEXP cpp_matrix_missing_row(SEXP x, double threshold, bool threshold_is_prop);
 extern "C" SEXP _cheapr_cpp_matrix_missing_row(SEXP x, SEXP threshold, SEXP threshold_is_prop) {
   BEGIN_CPP11
     return cpp11::as_sexp(cpp_matrix_missing_row(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<double>>(threshold), cpp11::as_cpp<cpp11::decay_t<bool>>(threshold_is_prop)));
   END_CPP11
 }
-// na_utils.cpp
+// nas.cpp
 SEXP cpp_matrix_missing_col(SEXP x, double threshold, bool threshold_is_prop);
 extern "C" SEXP _cheapr_cpp_matrix_missing_col(SEXP x, SEXP threshold, SEXP threshold_is_prop) {
   BEGIN_CPP11
@@ -146,17 +181,24 @@ extern "C" SEXP _cheapr_cpp_lead_sequence(SEXP size, SEXP k, SEXP partial) {
   END_CPP11
 }
 // utils.cpp
-R_xlen_t cpp_vector_size(SEXP x);
-extern "C" SEXP _cheapr_cpp_vector_size(SEXP x) {
+SEXP cpp_r_unnested_length(SEXP x);
+extern "C" SEXP _cheapr_cpp_r_unnested_length(SEXP x) {
   BEGIN_CPP11
-    return cpp11::as_sexp(cpp_vector_size(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+    return cpp11::as_sexp(cpp_r_unnested_length(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
   END_CPP11
 }
 // utils.cpp
-int cpp_vector_width(SEXP x);
-extern "C" SEXP _cheapr_cpp_vector_width(SEXP x) {
+SEXP cpp_lengths(SEXP x);
+extern "C" SEXP _cheapr_cpp_lengths(SEXP x) {
   BEGIN_CPP11
-    return cpp11::as_sexp(cpp_vector_width(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+    return cpp11::as_sexp(cpp_lengths(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+  END_CPP11
+}
+// utils.cpp
+SEXP cpp_new_list(R_xlen_t size, SEXP default_value);
+extern "C" SEXP _cheapr_cpp_new_list(SEXP size, SEXP default_value) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_new_list(cpp11::as_cpp<cpp11::decay_t<R_xlen_t>>(size), cpp11::as_cpp<cpp11::decay_t<SEXP>>(default_value)));
   END_CPP11
 }
 // which.cpp
@@ -169,25 +211,31 @@ extern "C" SEXP _cheapr_cpp_which_(SEXP x, SEXP invert) {
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
+    {"_cheapr_cpp_all_na",               (DL_FUNC) &_cheapr_cpp_all_na,               2},
+    {"_cheapr_cpp_any_na",               (DL_FUNC) &_cheapr_cpp_any_na,               1},
+    {"_cheapr_cpp_col_na_counts",        (DL_FUNC) &_cheapr_cpp_col_na_counts,        1},
     {"_cheapr_cpp_dbl_sequence",         (DL_FUNC) &_cheapr_cpp_dbl_sequence,         3},
     {"_cheapr_cpp_gcd",                  (DL_FUNC) &_cheapr_cpp_gcd,                  5},
     {"_cheapr_cpp_gcd2",                 (DL_FUNC) &_cheapr_cpp_gcd2,                 4},
     {"_cheapr_cpp_gcd2_vectorised",      (DL_FUNC) &_cheapr_cpp_gcd2_vectorised,      4},
     {"_cheapr_cpp_int_sequence",         (DL_FUNC) &_cheapr_cpp_int_sequence,         3},
+    {"_cheapr_cpp_is_na",                (DL_FUNC) &_cheapr_cpp_is_na,                1},
     {"_cheapr_cpp_lag_sequence",         (DL_FUNC) &_cheapr_cpp_lag_sequence,         3},
     {"_cheapr_cpp_lcm",                  (DL_FUNC) &_cheapr_cpp_lcm,                  3},
     {"_cheapr_cpp_lcm2",                 (DL_FUNC) &_cheapr_cpp_lcm2,                 4},
     {"_cheapr_cpp_lcm2_vectorised",      (DL_FUNC) &_cheapr_cpp_lcm2_vectorised,      4},
     {"_cheapr_cpp_lead_sequence",        (DL_FUNC) &_cheapr_cpp_lead_sequence,        3},
+    {"_cheapr_cpp_lengths",              (DL_FUNC) &_cheapr_cpp_lengths,              1},
     {"_cheapr_cpp_matrix_col_na_counts", (DL_FUNC) &_cheapr_cpp_matrix_col_na_counts, 1},
     {"_cheapr_cpp_matrix_missing_col",   (DL_FUNC) &_cheapr_cpp_matrix_missing_col,   3},
     {"_cheapr_cpp_matrix_missing_row",   (DL_FUNC) &_cheapr_cpp_matrix_missing_row,   3},
     {"_cheapr_cpp_matrix_row_na_counts", (DL_FUNC) &_cheapr_cpp_matrix_row_na_counts, 1},
+    {"_cheapr_cpp_missing_col",          (DL_FUNC) &_cheapr_cpp_missing_col,          3},
     {"_cheapr_cpp_missing_row",          (DL_FUNC) &_cheapr_cpp_missing_row,          3},
+    {"_cheapr_cpp_new_list",             (DL_FUNC) &_cheapr_cpp_new_list,             2},
     {"_cheapr_cpp_num_na",               (DL_FUNC) &_cheapr_cpp_num_na,               1},
+    {"_cheapr_cpp_r_unnested_length",    (DL_FUNC) &_cheapr_cpp_r_unnested_length,    1},
     {"_cheapr_cpp_row_na_counts",        (DL_FUNC) &_cheapr_cpp_row_na_counts,        1},
-    {"_cheapr_cpp_vector_size",          (DL_FUNC) &_cheapr_cpp_vector_size,          1},
-    {"_cheapr_cpp_vector_width",         (DL_FUNC) &_cheapr_cpp_vector_width,         1},
     {"_cheapr_cpp_which_",               (DL_FUNC) &_cheapr_cpp_which_,               2},
     {"_cheapr_cpp_which_na",             (DL_FUNC) &_cheapr_cpp_which_na,             1},
     {"_cheapr_cpp_which_not_na",         (DL_FUNC) &_cheapr_cpp_which_not_na,         1},